+++ src/main/java/koreatech/in/controller/UserController.java
@@ -19,8 +19,8 @@ import koreatech.in.domain.User.owner.Owner;
 import koreatech.in.dto.EmptyResponse;
 import koreatech.in.dto.ExceptionResponse;
 import koreatech.in.dto.RequestDataInvalidResponse;
-import koreatech.in.dto.global.TokenRefreshResponse;
-import koreatech.in.dto.global.auth.TokenRefreshRequest;
+import koreatech.in.dto.normal.auth.TokenRefreshResponse;
+import koreatech.in.dto.normal.auth.TokenRefreshRequest;
 import koreatech.in.dto.normal.user.request.AuthTokenRequest;
 import koreatech.in.dto.normal.user.request.CheckExistsEmailRequest;
 import koreatech.in.dto.normal.user.request.FindPasswordRequest;
