+++ src/main/java/koreatech/in/mapstruct/normal/auto/AuthConverter.java
@@ -2,8 +2,8 @@ package koreatech.in.mapstruct.normal.auto;
 
 import koreatech.in.domain.Auth.LoginResult;
 import koreatech.in.domain.Auth.RefreshToken;
-import koreatech.in.dto.global.TokenRefreshResponse;
-import koreatech.in.dto.global.auth.TokenRefreshRequest;
+import koreatech.in.dto.normal.auth.TokenRefreshResponse;
+import koreatech.in.dto.normal.auth.TokenRefreshRequest;
 import koreatech.in.dto.normal.user.response.LoginResponse;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
