+++ src/main/java/koreatech/in/service/UserServiceImpl.java
@@ -23,8 +23,8 @@ import koreatech.in.domain.User.User;
 import koreatech.in.domain.User.UserResponseType;
 import koreatech.in.domain.User.owner.Owner;
 import koreatech.in.domain.User.student.Student;
-import koreatech.in.dto.global.TokenRefreshResponse;
-import koreatech.in.dto.global.auth.TokenRefreshRequest;
+import koreatech.in.dto.normal.auth.TokenRefreshResponse;
+import koreatech.in.dto.normal.auth.TokenRefreshRequest;
 import koreatech.in.dto.normal.user.request.AuthTokenRequest;
 import koreatech.in.dto.normal.user.request.CheckExistsEmailRequest;
 import koreatech.in.dto.normal.user.request.FindPasswordRequest;
